requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloup==3.0.5
Cython==3.0.11
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
executing==2.1.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
glcontext==3.0.0
google-ai-generativelanguage==0.6.9
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
groq==0.11.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
ipython==8.27.0
isosurfaces==0.1.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.2
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.125
manim==0.18.1
ManimPango==0.6.0
mapbox_earcut==1.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
moderngl==5.11.1
moderngl-window==2.4.6
multidict==6.1.0
multipledispatch==1.0.0
mypy-extensions==1.0.0
narwhals==1.8.2
networkx==3.3
nltk==3.9.1
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.2
ptyprocess==0.7.0
pure_eval==0.2.3
# pyarrow==16.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.27.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pyglet==2.0.17
Pygments==2.18.0
# pyobjc-core==10.3.1
# pyobjc-framework-Cocoa==10.3.1
pyparsing==3.1.4
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rq==1.16.2
rsa==4.9
scipy==1.14.1
screeninfo==0.8.1
setuptools==75.1.0
six==1.16.0
skia-pathops==0.8.0.post1
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
srt==3.5.3
stack-data==0.6.3
streamlit==1.38.0
svgelements==1.9.6
tenacity==8.5.0
textblob==0.18.0.post0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
watchdog==4.0.2
wcwidth==0.2.13
wheel==0.44.0
yarl==1.11.1
